+++ packages/terra-framework-docs/src/terra-dev-site/test/popup/ArrowHorizontalAttachmentsPopup.test.jsx
@@ -66,6 +66,7 @@ class AlignmentPopup extends React.Component {
             isOpen={this.state.open}
             onRequestClose={this.handleRequestClose}
             targetRef={this.getButtonNode}
+            isHeaderDisabled={false}
           >
             <p className={cx('popup-text')}>This popup arrow was horizontally attached.</p>
           </Popup>