+++ packages/terra-framework-docs/src/terra-dev-site/test/popup/PopupWithContentWidth.test.jsx
@@ -100,6 +100,7 @@ class PopupWithContentWidth extends Component {
             isOpen={open}
             targetRef={this.getButtonNode}
             onRequestClose={this.handleRequestClose}
+            isHeaderDisabled={false}
           >
             <p>This is a popup with custom contentWidth and contentHeight</p>
           </Popup>