+++ packages/terra-framework-docs/src/terra-dev-site/test/popup/SizeSmallPopup.test.jsx
@@ -58,6 +58,7 @@ class DimensionPopup extends React.Component {
             isOpen={this.state.open}
             onRequestClose={this.handleRequestClose}
             targetRef={this.getButtonNode}
+            isHeaderDisabled={false}
           >
             <p>This popup is 40h by 160w.</p>
           </Popup>