+++ packages/terra-popup/src/Popup.jsx
@@ -122,7 +122,7 @@ const defaultProps = {
   contentWidth: '240',
   isArrowDisplayed: false,
   isContentFocusDisabled: false,
-  isHeaderDisabled: false,
+  isHeaderDisabled: true,
   isOpen: false,
   popupContentRole: 'dialog',
   isMenu: false,