+++ packages/terra-popup/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Disabled `close` button in popup by default.
+
 ## 6.84.0 - (May 20, 2024)
 
 * Changed